+++ src/main/java/kitchenpos/table/application/TableService.java
@@ -1,9 +1,12 @@
-package kitchenpos.order.application;
+package kitchenpos.table.application;
 
 import kitchenpos.exception.ErrorMessage;
-import kitchenpos.order.domain.*;
-import kitchenpos.order.dto.OrderTableRequest;
-import kitchenpos.order.dto.OrderTableResponse;
+import kitchenpos.table.domain.OrderEmpty;
+import kitchenpos.table.domain.OrderGuests;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.OrderTableRepository;
+import kitchenpos.table.dto.OrderTableRequest;
+import kitchenpos.table.dto.OrderTableResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -12,11 +15,11 @@ import java.util.List;
 @Service
 @Transactional
 public class TableService {
-    private final OrderRepository orderRepository;
+    private final TableValidator tableValidator;
     private final OrderTableRepository orderTableRepository;
 
-    public TableService(OrderRepository orderRepository, OrderTableRepository orderTableRepository) {
-        this.orderRepository = orderRepository;
+    public TableService(TableValidator tableValidator, OrderTableRepository orderTableRepository) {
+        this.tableValidator = tableValidator;
         this.orderTableRepository = orderTableRepository;
     }
 
@@ -31,9 +34,9 @@ public class TableService {
 
     public OrderTableResponse changeEmpty(Long orderTableId, OrderEmpty request) {
         OrderTable savedOrderTable = findOrderTableById(orderTableId);
-        List<Order> orders = findAllByOrderTableId(savedOrderTable.getId());
+        tableValidator.validateChangeEmpty(savedOrderTable);
 
-        savedOrderTable.changeEmpty(request.isEmpty(), orders);
+        savedOrderTable.changeEmpty(request.isEmpty());
 
         return OrderTableResponse.of(orderTableRepository.save(savedOrderTable));
     }
@@ -41,10 +44,6 @@ public class TableService {
     private OrderTable findOrderTableById(Long orderTableId) {
         return orderTableRepository.findById(orderTableId)
                 .orElseThrow(() -> new IllegalArgumentException(ErrorMessage.ORDER_TABLE_NOT_FOUND_BY_ID.getMessage()));
-    }
-
-    private List<Order> findAllByOrderTableId(Long orderTableId) {
-        return orderRepository.findAllByOrderTableId(orderTableId);
     }
 
     public OrderTableResponse changeNumberOfGuests(Long orderTableId, OrderGuests request) {
