+++ src/test/java/kitchenpos/table/application/TableGroupServiceTest.java
@@ -1,9 +1,9 @@
-package kitchenpos.order.application;
+package kitchenpos.table.application;
 
 import kitchenpos.exception.ErrorMessage;
-import kitchenpos.order.domain.*;
-import kitchenpos.order.dto.TableGroupRequest;
-import kitchenpos.order.dto.TableGroupResponse;
+import kitchenpos.table.domain.*;
+import kitchenpos.table.dto.TableGroupRequest;
+import kitchenpos.table.dto.TableGroupResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -15,13 +15,12 @@ import org.springframework.test.util.ReflectionTestUtils;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.ArgumentMatchers.*;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.Mockito.when;
 
 @DisplayName("단체 지정 비즈니스 테스트")
@@ -34,9 +33,6 @@ public class TableGroupServiceTest {
     @Mock
     private OrderTableRepository orderTableRepository;
 
-    @Mock
-    private OrderRepository orderRepository;
-
     @InjectMocks
     private TableGroupService tableGroupService;
 
@@ -45,6 +41,7 @@ public class TableGroupServiceTest {
     private OrderTable 두번째_주문_테이블;
     private OrderTable 세번째_주문_테이블;
     private OrderTable 네번째_주문_테이블;
+    private OrderTable 다섯번째_주문_테이블;
     private List<OrderTable> 주문_테이블_목록;
 
     @BeforeEach
@@ -55,15 +52,16 @@ public class TableGroupServiceTest {
         두번째_주문_테이블 = new OrderTable(4, true);
         세번째_주문_테이블 = new OrderTable(2, true);
         네번째_주문_테이블 = new OrderTable(2, true);
+        다섯번째_주문_테이블 = new OrderTable(2, false);
 
         ReflectionTestUtils.setField(우아한형제들_단체그룹, "id", 1L);
         ReflectionTestUtils.setField(첫번째_주문_테이블, "id", 1L);
         ReflectionTestUtils.setField(두번째_주문_테이블, "id", 2L);
         ReflectionTestUtils.setField(세번째_주문_테이블, "id", 3L);
         ReflectionTestUtils.setField(네번째_주문_테이블, "id", 4L);
+        ReflectionTestUtils.setField(다섯번째_주문_테이블, "id", 5L);
 
         주문_테이블_목록 = Arrays.asList(첫번째_주문_테이블, 두번째_주문_테이블);
-        우아한형제들_단체그룹.group(주문_테이블_목록);
     }
 
     @DisplayName("주문 테이블들이 2개 이상 있어야 한다.")
@@ -98,9 +96,9 @@ public class TableGroupServiceTest {
     @Test
     void 주문_테이블들은_빈_테이블이어야_한다() {
         // given
-        when(orderTableRepository.findAllById(anyList())).thenReturn(Arrays.asList(첫번째_주문_테이블, 두번째_주문_테이블));
+        when(orderTableRepository.findAllById(anyList())).thenReturn(Arrays.asList(첫번째_주문_테이블, 다섯번째_주문_테이블));
 
-        TableGroupRequest tableGroupRequest = new TableGroupRequest(Arrays.asList(첫번째_주문_테이블.getId(), 두번째_주문_테이블.getId()));
+        TableGroupRequest tableGroupRequest = new TableGroupRequest(Arrays.asList(첫번째_주문_테이블.getId(), 다섯번째_주문_테이블.getId()));
 
         // when, then
         assertThatIllegalArgumentException()
@@ -119,6 +117,8 @@ public class TableGroupServiceTest {
                 .thenReturn(Arrays.asList(첫번째_주문_테이블, 두번째_주문_테이블));
 
         TableGroupRequest tableGroupRequest = new TableGroupRequest(Arrays.asList(첫번째_주문_테이블.getId(), 두번째_주문_테이블.getId()));
+
+        우아한형제들_단체그룹.group(주문_테이블_목록);
 
         // when, then
         assertThatIllegalArgumentException()
@@ -144,33 +144,5 @@ public class TableGroupServiceTest {
             assertEquals(response.getId(), 우아한형제들_단체그룹.getId());
             assertEquals(response.getOrderTables().size(), 우아한형제들_단체그룹.getOrderTables().size());
         });
-    }
-
-    @DisplayName("단체 내 주문 테이블들의 상태는 조리 중이거나 식사중이면 안된다.")
-    @Test
-    void 단체_내_주문_테이블들의_상태는_조리_중이거나_식사중이면_안된다() {
-        // given
-        when(tableGroupRepository.findById(anyLong())).thenReturn(Optional.of(우아한형제들_단체그룹));
-        when(orderRepository.findAllByOrderTableIds(anyList()))
-                .thenReturn(Arrays.asList(new Order(첫번째_주문_테이블, OrderStatus.COMPLETION), new Order(두번째_주문_테이블, OrderStatus.COOKING)));
-
-        // when, then
-        assertThatIllegalArgumentException()
-                .isThrownBy(() -> tableGroupService.ungroup(우아한형제들_단체그룹.getId()))
-                .withMessage(ErrorMessage.ORDER_CANNOT_BE_CHANGED.getMessage());
-    }
-
-    @DisplayName("단체 지정을 삭제할 수 있다.")
-    @Test
-    void 단체_지정을_삭제할_수_있다() {
-        // given
-        when(tableGroupRepository.findById(anyLong())).thenReturn(Optional.of(우아한형제들_단체그룹));
-        when(orderRepository.findAllByOrderTableIds(anyList()))
-                .thenReturn(Arrays.asList(new Order(첫번째_주문_테이블, OrderStatus.COMPLETION), new Order(두번째_주문_테이블, OrderStatus.COMPLETION)));
-        // when
-        tableGroupService.ungroup(우아한형제들_단체그룹.getId());
-
-        // then
-        assertTrue(주문_테이블_목록.stream().allMatch(orderTable -> orderTable.getTableGroup() == null));
     }
 }
